+++ src/App.tsx
@@ -1,17 +1,23 @@
-import React from 'react';
-import { BrowserRouter as Router } from 'react-router-dom';
-import RoutesComponent from './Pages/Routes';
-import Globalstyle from './Styles/GlobalStyle';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Layout from './Layout';
+import Intro from './Pages/Intro';
+import Login from './Pages/Intro/Login';
+import HyeyeonPage from './Pages/hyeyeon';
+import SoyoonPage from './Pages/soyoon';
 
-function App(): JSX.Element {
+const App = () => {
   return (
-    <>
-      <Globalstyle />
-      <Router>
-        <RoutesComponent />
-      </Router>
-    </>
+    <BrowserRouter>
+      <Routes>
+        <Route element={<Layout />}>
+          <Route path="/" element={<Intro />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/hyeyeon" element={<HyeyeonPage />} />
+          <Route path="/soyoon" element={<SoyoonPage />} />
+        </Route>
+      </Routes>
+    </BrowserRouter>
   );
-}
+};
 
 export default App;
