--- src/Components/Common/Layout.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-interface Props {
-  children: React.ReactNode;
-}
-
-const Layout: React.FC<Props> = (props: Props) => {
-  return <div>{props.children}</div>;
-};
-
-export default Layout;
