+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
-import Routes from './Pages/Routes';
+import App from './App';
+import GlobalStyle from './Styles/GlobalStyle';
 import './Styles/common.scss';
 import './Styles/reset.scss';
 
@@ -10,9 +10,8 @@ const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
   <React.StrictMode>
     <RecoilRoot>
-      <BrowserRouter>
-        <Routes />
-      </BrowserRouter>
+      <GlobalStyle />
+      <App />
     </RecoilRoot>
   </React.StrictMode>
 );
